refactor: move block selection note linking into host
